requirements.txt
pystray
Pillow
py7zr
